:wrench: code cleanup
