+++ app/frontend/src/pages/menu/Menu.module.css
@@ -3,9 +3,9 @@
     display: flex;
     flex-direction: column;
     margin-top: 20px;
-    max-width: 90em;
-    margin-right: 5em;
-    margin-left: 5em;
+    max-width: 80%;
+    margin-right: 10%;
+    margin-left: 10%;
 }
 .rowheader {
     flex: 1;
@@ -20,19 +20,17 @@
     padding-top: 10px;
     width: 100%;
     height: auto;
-    flex: 1;
-    display: flex;
-    overflow: auto;
-    flex-wrap: wrap;
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+    gap: 10px;
     font-size: var(--fontSizeBase400);
 }
 .box {
-    flex: 1;
+    flex: 1 1 250px;
     text-align: center;
     margin-top: 0px;
     overflow: hidden;
     text-overflow: ellipsis;
-    min-width: 200px;
     height: 100px;
     border-radius: 20px;
     margin-right: 10px;
@@ -59,20 +57,4 @@
     top: 45%;
     align-self: center;
     overflow-wrap: break-word;
-}
-
-.dialog {
-    padding: 0 !important;
-    min-width: 1100px;
-    line-height: 1.5;
-}
-
-@media (max-width: 650px) {
-    .dialog {
-        min-width: 400px;
-    }
-}
-
-.dialogContent {
-    padding: 20px;
 }
